requirements.txt
appdirs
APScheduler
dill
cryptography
cursor
readchar
colorama
infi.systray
PySide6
maybe-else
pandas
pathmagic
pymiscutils
pysubtypes
typepy
